[KYUUBI #2604] Hive Backend Engine - Multi tenancy support
